Refactor employee dashboard #226
